feat: integrate leave approval with attendance updates and fix half-day leave calculation handling

+++ backend/app/schemas/leave_structure.py
@@ -127,7 +127,7 @@ class LeaveAssignmentOut(BaseModel):
 # ─────────────────────────────────────────────────────────────
 
 class LeaveBalanceSummaryOut(BaseModel):
-    allocated: int
+    allocated: float
     used:      float
     remaining: float
     excess:    float
